move another non-class method
