+++ main/provider/helpers.ts
@@ -32,6 +32,46 @@ function getRate (address: Address) {
  
   return (rate || { usd: { price: 0 } }) as Rate
 }
+
+export function checkExistingNonceGas (tx: TransactionData) {
+  const { from, nonce } = tx
+
+  const reqs = store('main.accounts', from, 'requests')
+  const requests = Object.keys(reqs || {}).map(key => reqs[key])
+  const existing = requests.filter(r => (
+    r.mode === 'monitor' && 
+    r.status !== 'error' && 
+    r.data.nonce === nonce
+  ))
+
+  if (existing.length > 0) {
+    if (tx.maxPriorityFeePerGas && tx.maxFeePerGas) {
+      const existingFee = Math.max(...existing.map(r => r.data.maxPriorityFeePerGas))
+      const existingMax = Math.max(...existing.map(r => r.data.maxFeePerGas))
+      const feeInt = parseInt(tx.maxPriorityFeePerGas)
+      const maxInt = parseInt(tx.maxFeePerGas)
+      if (existingFee * 1.1 >= feeInt || existingMax * 1.1 >= maxInt) {
+        // Bump fees by 10%
+        const bumpedFee = Math.max(Math.ceil(existingFee * 1.1), feeInt)
+        const bumpedBase = Math.max(Math.ceil((existingMax - existingFee) * 1.1), Math.ceil(maxInt - feeInt))
+        tx.maxFeePerGas = '0x' + (bumpedBase + bumpedFee).toString(16)
+        tx.maxPriorityFeePerGas = '0x' + bumpedFee.toString(16)
+        tx.feesUpdated = true
+      }
+    } else if (tx.gasPrice) {
+      const existingPrice = Math.max(...existing.map(r => r.data.gasPrice))
+      const priceInt = parseInt(tx.gasPrice)
+      if (existingPrice >= priceInt) {
+        // Bump price by 10%
+        const bumpedPrice = Math.ceil(existingPrice * 1.1)
+        tx.gasPrice = '0x' + bumpedPrice.toString(16)
+        tx.feesUpdated = true
+      }
+    }
+  }
+
+  return tx
+}
   
 export function loadAssets (accountId: string) {
   const balances: Balance[] = store('main.balances', accountId) || []
